Implemented Darwin sampling logic
Added Darwin section in sample.c, which suspends and resumes each mach thread and gets its state directly rather than relying on signal handlers in Linux. 

Also changed os_dep functions for both Linux and Darwin. The array of thread_info is now put on the heap, managed by the create and destroy function. 

For the Swift Profile Recorderconverter a bug has been fixed in LLVMJSONOutputParserHandler, although we are still failing to output the functions in the shared cache, as shown in the flame graph.