shorten thread name to fit Linux's 16 char limit (#28)
Whilst Profile Recorder is sampling the process, it renames the thread
that controls the sampling. Currently it set it to
`ProfileRecorder-sampling` but Linux has a 16 character limit for thread
names, so it gets truncated.

Let's name it `swipr-sampling` (swipr for SWIft Profile Recorder)
instead.

Please note, the intent is that nobody ever even sees this thread name
because it won't appear in the samples (it filters out its own thread)
and unless it's actively sampling it won't set that thread name. So the
only time when you'd see this thread name is if you either sampled the
process with an external profiler whilst also concurrently sampling with
Profile Recorder at the same time. Or of course, in case Profile
Recorder has a bug and crashes the process.

Co-authored-by: Johannes Weiss <johannes@jweiss.io>